correct error with Polygon / ExPolygon add to cmakelist add needed function (mostly copy from slic3rPE)

+++ xs/src/libslic3r/ExPolygon.cpp
@@ -191,16 +191,16 @@ void
 ExPolygon::remove_point_too_near(const coord_t tolerance) {
     size_t id = 1;
     while (id < this->contour.points.size() - 1) {
-        size_t newdist = min(this->contour.points[id].distance_to(this->contour.points[id - 1])
+        double newdist = min(this->contour.points[id].distance_to(this->contour.points[id - 1])
             , this->contour.points[id].distance_to(this->contour.points[id + 1]));
-        if (newdist < tolerance) {
+        if (newdist < (double)tolerance) {
             this->contour.points.erase(this->contour.points.begin() + id);
             newdist = this->contour.points[id].distance_to(this->contour.points[id - 1]);
         }
         //go to next one
         //if you removed a point, it check if the next one isn't too near from the previous one.
         // if not, it byepass it.
-        if (newdist > tolerance) {
+        if (newdist > (double)tolerance) {
             ++id;
         }
     }
